2005-10-04 20:06 UTC+0200 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * xharbour/include/hbdefs.h   * xharbour/source/common/hbver.c   * xharbour/source/compiler/genc.c   * xharbour/source/ct/charone.c   * xharbour/source/rtl/filesys.c   * xharbour/source/rtl/hbserv.c   * xharbour/source/rtl/gtstd/gtstd.c    * cleaned some BCC warnings
  * xharbour/source/rdd/dbcmd.c
  * xharbour/source/rdd/rddsys.prg
   ! fixed default RDD setting

+++ xharbour/source/rtl/filesys.c
@@ -1,5 +1,5 @@
 /*
- * $Id: filesys.c,v 1.150 2005/09/19 23:21:29 druzus Exp $
+ * $Id: filesys.c,v 1.151 2005/10/02 12:35:11 druzus Exp $
  */
 
 /*
@@ -3026,7 +3026,7 @@ USHORT HB_EXPORT  hb_fsChDrv( BYTE nDrive )
       HB_FS_SETDRIVE( nDrive );
       HB_FS_GETDRIVE( uiNewDrive );
 
-      if ( nDrive == uiNewDrive )
+      if( ( UINT ) nDrive == uiNewDrive )
       {
          uiResult = 0;
          hb_fsSetError( 0 );
@@ -3081,7 +3081,7 @@ USHORT HB_EXPORT  hb_fsIsDrv( BYTE nDrive )
       HB_FS_GETDRIVE( uiSave );
       HB_FS_SETDRIVE( nDrive );
       HB_FS_GETDRIVE( uiNewDrive );
-      if( nDrive != uiNewDrive )
+      if( ( UINT ) nDrive != uiNewDrive )
       {
          uiResult = (USHORT) FS_ERROR;
          hb_fsSetError( (USHORT) FS_ERROR );
